Continued working on serializers and REST. Added AJAX requests.
