+++ siteweather/static/js/previous_page_redirection.js
@@ -1,0 +1,6 @@
+$(document).ready(function () {
+  $("#previous_page").click(function (event) {
+    event.preventDefault();
+    location.href = document.referrer;
+  })
+})
