+++ siteweather/templates/profile/password_update.html
@@ -8,12 +8,9 @@
   {% include 'inc/_sidebar.html' %}
 {% endblock %}
 
-{% block pagination %}
-{% endblock %}
-
 {% block content %}
   {% if request.user.is_authenticated %}
-    <form action="" method="post" autocomplete="off" enctype="multipart/form-data">
+    <form id="password-update-form" autocomplete="off" enctype="multipart/form-data">
       {% csrf_token %}
       <div class="card mb-4">
         <div class="card-header">
@@ -67,4 +64,29 @@
 
 {% block time %}
   {% include 'inc/_time.html' %}
+{% endblock %}
+
+{% block javascript %}
+  <script>
+    $(document).ready(function () {
+      let token = '{{ csrf_token }}';
+      $("#password-update-form").submit(function (event) {
+        event.preventDefault();
+        let serializerData = $(this).serialize();
+        $.ajax({
+          type: 'PUT',
+          data: serializerData,
+          headers: {"X-CSRFToken": token},
+          async: false,
+          success: function (response) {
+            location.href = '{% url 'siteweather:profile:profile' pk %}';
+            alert(response.message);
+          },
+          error: function (response) {
+            alert(response["responseJSON"]["error"]);
+          }
+        })
+      })
+    })
+  </script>
 {% endblock %}
