+++ siteweather/templates/siteweather/find_by.html
@@ -1,4 +1,5 @@
 {% extends 'base.html' %}
+{% load static %}
 
 {% block title %}
   Find by city name
@@ -6,16 +7,6 @@
 
 {% block sidebar %}
   {% include 'inc/_sidebar.html' %}
-{% endblock %}
-
-{% block pagination %}
-  {% if request.META.HTTP_REFERER|cut:"http://127.0.0.1:8000" != request.get_full_path %}
-    {% if request.META.HTTP_REFERER is not None %}
-      <a href="{{ request.META.HTTP_REFERER }}" class="btn btn-outline-primary btn-block">
-        Previous page
-      </a>
-    {% endif %}
-  {% endif %}
 {% endblock %}
 
 {% block content %}
@@ -28,7 +19,7 @@
             City name:
           </label>
           <input type="text" class="form-control" name="city_name"
-                   value="{{ data.username }}" required="" id="id_city_name">
+                 value="{{ data.username }}" required="" id="id_city_name">
           {% if errors %}
             <br>
             {% for error in errors.city_name %}
@@ -48,4 +39,7 @@
 
 {% block time %}
   {% include 'inc/_time.html' %}
+{% endblock %}
+
+{% block javascript %}
 {% endblock %}
