+++ siteweather/templates/siteweather/home.html
@@ -49,7 +49,7 @@
       </div>
     </div>
   {% endfor %}
-  <div class="card card-header">
+  <div class="card card-header" id="content">
     <div class="text-muted">
       Filter
     </div>
@@ -75,4 +75,12 @@
 
 {% block time %}
   {% include 'inc/_time.html' %}
+{% endblock %}
+
+{% block javascript %}
+  <script>
+    $(document).ready(function () {
+
+    })
+  </script>
 {% endblock %}
