+++ task/urls.py
@@ -12,6 +12,7 @@ urlpatterns = [
     path('docs/', schema_view, name="docs"),
     path('admin/logout/', AdminLogoutView.as_view()),
     path('admin/', admin.site.urls),
+    # path('', admin.site.urls),
     path('', include('siteweather.urls', namespace='siteweather')),
     path('api-auth/', include('rest_framework.urls'))
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
