+++ templates/inc/_pagination.html
@@ -3,14 +3,14 @@
     {% if previous %}
       <li class="page-item">
         <a class="page-link" href="{{ previous }}">
-          ‹‹
+          Previous elements
         </a>
       </li>
     {% endif %}
     {% if next %}
       <li class="page-item">
         <a class="page-link" href="{{ next }}">
-          ››
+          Next elements
         </a>
       </li>
     {% endif %}
