+++ templates/inc/_users_pagination.html
@@ -3,14 +3,14 @@
     {% if previous %}
       <li class="page-item">
         <a class="page-link" href="{{ previous }}">
-          ‹‹
+          Previous users
         </a>
       </li>
     {% endif %}
     {% if next %}
       <li class="page-item">
         <a class="page-link" href="{{ next }}">
-          ››
+          Next users
         </a>
       </li>
     {% endif %}
